bring in algo-order adapter, enable staging urls
this pr brings in the exchange adapter for our only exchange
bitfinex, and fixes places where urls to the apis where not
dynamic, but relied on defaults in the abstractions.

the `lib/ws_servers/algos/ao_adapter.js` file is coming from
`bfx-hf-ext-plugin-bitfinex`.

with that PR, devs can run algo orders easily on staging, without
changing it in the submodules, i.e. the
`bfx-hf-ext-plugin-bitfinex` adapter

with just one exchange the exchange pool is superfluous, and will
get removed in next refactors.

the next step for simplifying the repos is to simplify algo orders,
e.g. removing one of the current two used algo servers (one inside
repo, one via a submodule called `bfx-hf-algo-server`)